+++ src/snowflake/cli/_plugins/nativeapp/codegen/sandbox.py
@@ -24,6 +24,7 @@ from typing import Any, Mapping, Optional, Sequence, Union
 from venv import EnvBuilder
 
 from click.exceptions import ClickException
+from snowflake.cli.api.config import get_subprocess_encoding
 
 EnvVars = Mapping[str, str]  # Only support str -> str for cross-platform compatibility
 
@@ -70,10 +71,12 @@ def _execute_python_interpreter(
     else:
         args = [arg for arg in python_executable]
     args.append("-")
+    encoding = get_subprocess_encoding()
     return subprocess.run(
         args,
         capture_output=True,
         text=True,
+        encoding=encoding,
         input=script_source,
         timeout=timeout,
         cwd=cwd,